Move DeltaLengthByteArray/DeltaByteArray decoders to separate file
