+++ extension/parquet/decoder/CMakeLists.txt
@@ -2,6 +2,7 @@ add_library_unity(
   duckdb_parquet_decoders
   OBJECT
   delta_binary_packed_decoder.cpp
+  delta_byte_array_decoder.cpp
   dictionary_decoder.cpp
   rle_decoder.cpp)
 set(PARQUET_EXTENSION_FILES
