+++ extension/parquet/include/reader/string_column_reader.hpp
@@ -29,13 +29,8 @@ public:
 	StringColumnReader(ParquetReader &reader, LogicalType type_p, const SchemaElement &schema_p, idx_t schema_idx_p,
 	                   idx_t max_define_p, idx_t max_repeat_p);
 	idx_t fixed_width_string_length;
-	idx_t delta_offset = 0;
 
 public:
-	void PrepareDeltaLengthByteArray(ResizeableBuffer &buffer) override;
-	void PrepareDeltaByteArray(ResizeableBuffer &buffer) override;
-	void DeltaByteArray(uint8_t *defines, idx_t num_values, parquet_filter_t &filter, idx_t result_offset,
-	                    Vector &result) override;
 	static uint32_t VerifyString(const char *str_data, uint32_t str_len, const bool isVarchar);
 	uint32_t VerifyString(const char *str_data, uint32_t str_len);
